xử lý lưu thông tin lệnh vào DB

+++ ObjectInfo/Msg_Tprl_Repo_Detail_Info.cs
@@ -9,10 +9,10 @@
         public decimal Execqty { get; set; }
         public decimal Execpx { get; set; }
         public decimal Price { get; set; }
-        public decimal Reposinterest { get; set; }
-        public decimal Hedgerate { get; set; }
-        public decimal Settlvalue2 { get; set; }
-        public decimal Settlvalue { get; set; }
+        public double Reposinterest { get; set; }
+        public double Hedgerate { get; set; }
+        public double Settlvalue2 { get; set; }
+        public double Settlvalue { get; set; }
         public decimal Price2 { get; set; }
         public string Remark { get; set; } = string.Empty;
         public string Lastchange { get; set; } = string.Empty;
